Docker/backend-app/requirements.txt
## Python dependencies for backend app
flask==3.0.3
gunicorn==23.0.0
psycopg2-binary==2.9.9